requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
anyio==4.8.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.28.1

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# MCP integration
anthropic==0.40.0
openai==1.54.0

# Additional utilities
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.1.1
pyjwt==2.8.0
